+++ src/main/java/dev/orne/beans/BeanValidationUtils.java
@@ -32,6 +32,8 @@ import javax.validation.Validator;
 
 import org.apache.commons.lang3.Validate;
 
+import dev.orne.beans.ValidBeanReference.ValidBeanReferenceValidator;
+
 /**
  * Utility class for bean validations.
  * 
@@ -92,6 +94,21 @@ public final class BeanValidationUtils {
             @Nonnull
             final Class<?>... groups) {
         return validate(obj, groups).isEmpty();
+    }
+
+    /**
+     * Validates if the specified bean is a valid bean reference.
+     * 
+     * @param obj The bean to validate
+     * @return If the object is a valid bean reference
+     * @throws IllegalArgumentException if object is {@code null}
+     * @throws ValidationException if a non recoverable error happens
+     *         during the validation process
+     */
+    public static boolean isValidBeanReference(
+            @Nonnull
+            final Object obj) {
+        return ValidBeanReferenceValidator.isValid(obj);
     }
 
     /**